requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
idna==3.4
jeepney==0.8.0
mss==7.0.1
mysqlclient==2.1.1
Pillow==9.3.0
pyscreenshot==3.0
pytz==2022.6
requests==2.28.1
schedule==1.1.0
sqlparse==0.4.3
urllib3==1.26.13
